fix: make & apply custom d.ts
커스텀 d.ts 파일을 만든 후, import 구문에 일괄 적용함.
