+++ src/component/TodoItem/EmptyTodoItem.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {FunctionComponent} from 'react';
-import {Todo} from "../todoTypes";
 import {EmptyTodoItemStyle} from './TodoItem.style';
+import {Todo} from "custom-types";
 
 export const EmptyTodoItem: FunctionComponent<Todo.EmptyTodoItemProp> = function () {
     const content = "todo list is empty. add your todo item...";
