+++ src/component/TodoList/TodoList.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import {FunctionComponent} from 'react';
-import {Todo} from "../todoTypes";
 import {TodoItem} from "../TodoItem/TodoItem";
 import TodoListStyle from "./TodoList.style";
 import {EmptyTodoItem} from "../TodoItem/EmptyTodoItem";
+import {Todo} from "custom-types";
 
 export const TodoList: FunctionComponent<Todo.TodoListProp> = function ({todoInfos, deleteTodoItem, toggleDone}) {
     const todoItems = todoInfos.map((info, i) => <TodoItem
